Move FWFR into a separate package
Just a precursor to doing any work on it.
